Always pass --lib to cargo with link-args
Cargo will throw an error when the target of the compilation is
ambiguous. This leads to an error when there's an additional main.rs,
e.g. for easy testing during development. Specifying `--lib` makes it
clear that we only want to build the library.

Before

```
~/.cargo/bin/cargo rustc --release -- -C link-args=-Wl,-undefined,dynamic_lookup
error: extra arguments to `rustc` can only be passed to one target,
consider filtering
```

After

```
~/.cargo/bin/cargo rustc --release --lib -- -C link-args=-Wl,-undefined,dynamic_lookup
```